+++ http/rest_tools/src/main/kotlin/com/hexagonkt/http/rest/tools/Http.kt
@@ -1,4 +1,4 @@
-package com.hexagonkt.http.test
+package com.hexagonkt.http.rest.tools
 
 import com.hexagonkt.core.media.MediaType
 import com.hexagonkt.core.urlOf
