+++ http/rest_tools/src/main/kotlin/com/hexagonkt/http/rest/tools/openapi/OpenApiHandler.kt
@@ -1,4 +1,4 @@
-package com.hexagonkt.http.test.openapi
+package com.hexagonkt.http.rest.tools.openapi
 
 import com.hexagonkt.core.fail
 import com.hexagonkt.core.require
